Update entity data in-memory and write files only once
